+++ tests/testthat/test-sum_time.R
@@ -10,8 +10,10 @@ test_that("sum_time() | linear test", {
                           lubridate::as.interval(lubridate::dhours(1),
                                                  as.Date("1970-01-01")),
                           cycle = NULL,
+                          reverse = FALSE,
                           na_rm = FALSE),
                  hms::hms(as.numeric(lubridate::dhours(30)))) # 30:00:00
+
     expect_equal(sum_time(hms::as_hms(NA), # !
                           c(lubridate::dhours(1), lubridate::dminutes(30)),
                           lubridate::hours(1),
@@ -24,8 +26,10 @@ test_that("sum_time() | linear test", {
                           lubridate::as.interval(lubridate::dhours(1),
                                                  as.Date("1970-01-01")),
                           cycle = NULL,
+                          reverse = FALSE,
                           na_rm = FALSE), # !
                  hms::as_hms(NA))
+
     expect_equal(sum_time(hms::as_hms(NA), # !
                           c(lubridate::dhours(1), lubridate::dminutes(30)),
                           lubridate::hours(1),
@@ -38,6 +42,7 @@ test_that("sum_time() | linear test", {
                           lubridate::as.interval(lubridate::dhours(1),
                                                  as.Date("1970-01-01")),
                           cycle = NULL,
+                          reverse = FALSE,
                           na_rm = TRUE), # !
                  hms::hms(as.numeric(lubridate::dhours(30)))) # 30:00:00
 })
@@ -54,8 +59,10 @@ test_that("sum_time() | circular test", {
                           lubridate::as.interval(lubridate::dhours(1),
                                                  as.Date("1970-01-01")),
                           cycle = lubridate::ddays(), # !
+                          reverse = FALSE,
                           na_rm = FALSE),
                  hms::parse_hm("06:00")) # 06:00 | 30 - 24
+
     expect_equal(sum_time(hms::as_hms(NA),
                           c(lubridate::dhours(1), lubridate::dminutes(30)),
                           lubridate::hours(1),
@@ -68,8 +75,23 @@ test_that("sum_time() | circular test", {
                           lubridate::as.interval(lubridate::dhours(1),
                                                  as.Date("1970-01-01")),
                           cycle = lubridate::ddays(),
+                          reverse = FALSE,
                           na_rm = TRUE),
                  hms::parse_hm("06:00"))
+
+    expect_equal(sum_time(c(lubridate::dhours(-1), lubridate::dhours(-2)),
+                          c(lubridate::hours(-1), lubridate::hours(-21)),
+                          cycle = lubridate::ddays(),
+                          reverse = FALSE,
+                          na_rm = FALSE),
+                 - hms::parse_hm("01:00"))
+
+    expect_equal(sum_time(c(lubridate::dhours(-1), lubridate::dhours(-2)),
+                          c(lubridate::hours(-1), lubridate::hours(-21)),
+                          cycle = lubridate::ddays(),
+                          reverse = TRUE,
+                          na_rm = FALSE),
+                 hms::parse_hm("23:00"))
 })
 
 test_that("vct_sum_time()| linear test", {
@@ -90,9 +112,11 @@ test_that("vct_sum_time()| linear test", {
                                 lubridate::as.interval(lubridate::dhours(1),
                                                        as.Date("1970-01-01"))),
                               cycle = NULL,
+                              reverse = FALSE,
                               na_rm = FALSE),
                  c(hms::hms(108000),
                    hms::hms(61200))) # 30:00:00 | 17:00:00
+
     expect_equal(vct_sum_time(c(hms::as_hms(NA), hms::as_hms(NA)), # !
                               c(lubridate::dhours(1), lubridate::dminutes(30)),
                               c(lubridate::hours(1), lubridate::hours(1)),
@@ -111,8 +135,10 @@ test_that("vct_sum_time()| linear test", {
                                 lubridate::as.interval(lubridate::dhours(1),
                                                        as.Date("1970-01-01"))),
                               cycle = NULL,
+                              reverse = FALSE,
                               na_rm = FALSE), # !
                  c(hms::as_hms(NA), hms::as_hms(NA)))
+
     expect_equal(vct_sum_time(c(hms::as_hms(NA), hms::as_hms(NA)), # !
                               c(lubridate::dhours(1), lubridate::dminutes(30)),
                               c(lubridate::hours(1), lubridate::hours(1)),
@@ -131,6 +157,7 @@ test_that("vct_sum_time()| linear test", {
                                 lubridate::as.interval(lubridate::dhours(1),
                                                        as.Date("1970-01-01"))),
                               cycle = NULL,
+                              reverse = FALSE,
                               na_rm = TRUE), # !
                  c(hms::hms(108000),
                    hms::hms(61200))) # 30:00:00 | 17:00:00
@@ -154,9 +181,11 @@ test_that("vct_sum_time()| circular test", {
                                 lubridate::as.interval(lubridate::dhours(1),
                                                        as.Date("1970-01-01"))),
                               cycle = lubridate::ddays(), # !
+                              reverse = FALSE,
                               na_rm = FALSE),
                  c(hms::parse_hm("06:00"),
                    hms::parse_hm("17:00"))) # 06:00:00 | 17:00:00
+
     expect_equal(vct_sum_time(c(hms::as_hms(NA), hms::as_hms(NA)),
                               c(lubridate::dhours(1), lubridate::dminutes(30)),
                               c(lubridate::hours(1), lubridate::hours(1)),
@@ -175,24 +204,40 @@ test_that("vct_sum_time()| circular test", {
                                 lubridate::as.interval(lubridate::dhours(1),
                                                        as.Date("1970-01-01"))),
                               cycle = lubridate::ddays(), # !
+                              reverse = FALSE,
                               na_rm = TRUE),
                  c(hms::hms(21600), hms::hms(61200)))
+
+    expect_equal(vct_sum_time(c(lubridate::dhours(-24), lubridate::dhours(-1)),
+                              c(lubridate::hours(-6), lubridate::hours(-48)),
+                              cycle = lubridate::ddays(),
+                              reverse = FALSE,
+                              na_rm = FALSE),
+                 c(- hms::parse_hm("06:00"), - hms::parse_hm("01:00")))
+
+    expect_equal(vct_sum_time(c(lubridate::dhours(-24), lubridate::dhours(-1)),
+                              c(lubridate::hours(-6), lubridate::hours(-48)),
+                              cycle = lubridate::ddays(),
+                              reverse = TRUE,
+                              na_rm = FALSE),
+                 c(hms::parse_hm("18:00"), hms::parse_hm("23:00")))
 })
 
-test_that("build_sum() | error test", {
-    expect_error(build_sum(1, vectorize = FALSE, cycle = NULL, na_rm = FALSE),
+test_that("sum_time_build() | error test", {
+    expect_error(sum_time_build(1, vectorize = FALSE, cycle = NULL,
+                                na_rm = FALSE),
                  "Assertion on 'x' failed")
-    expect_error(build_sum(hms::hms(1), vectorize = "", cycle = NULL,
+    expect_error(sum_time_build(hms::hms(1), vectorize = "", cycle = NULL,
                            na_rm = TRUE),
                  "Assertion on 'vectorize' failed")
-    expect_error(build_sum(hms::hms(1), vectorize = FALSE, cycle = "",
+    expect_error(sum_time_build(hms::hms(1), vectorize = FALSE, cycle = "",
                            na_rm = FALSE),
                  "Assertion on 'cycle' failed")
-    expect_error(build_sum(hms::hms(1), vectorize = FALSE, cycle = NULL,
+    expect_error(sum_time_build(hms::hms(1), vectorize = FALSE, cycle = NULL,
                            na_rm = ""),
                  "Assertion on 'na_rm' failed")
 
-    expect_error(build_sum(hms::hms(1), c(hms::hms(1), hms::hms(1)),
+    expect_error(sum_time_build(hms::hms(1), c(hms::hms(1), hms::hms(1)),
                            vectorize = TRUE),
                  "All values in '...' must ")
 })